+++ Stitch/Graph/StitchAI/Mapping/CodeToSyntax/CodeToStitchPatchData.swift
@@ -208,16 +208,6 @@ extension SwiftParserPatchData {
             case .binding(let refName):
                 return .expression(.ref(refName))
                 
-                // TODO: get this working for edge logic
-                //                SwiftParserPatchData
-                //                    .processIncomingConnectionData(upstreamCoordinate: upstreamCoordinate,
-                //                                                   downstreamCoordinate: .init(node_id: patchNodeData.id,
-                //                                                                               port_index: portIndex),
-                //                                                   nativePatchNodes: nativePatchNodes,
-                //                                                   nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-                //                                                   patchConnections: &patchConnections)
-                
-                
             case .subscriptRef(let subscriptRef):
                 // Recursively call data
                 guard let result = try SwiftParserInitializerType.subscriptRef(subscriptRef)
@@ -230,47 +220,6 @@ extension SwiftParserPatchData {
                 
             case .value(let argType):
                 return .expression(.portValuesInit([argType]))
-//                let portDataList: [PortValueCodeType] = try argType.derivePortValues().map { portData in
-//                    switch portData {
-//                    case .value(let portValue):
-//                        return .value(.init(value: portValue.value,
-//                                         value_type: portValue.value_type))
-//                        
-//                        // TODO: custom node type
-//                        // Update node's custom value type if relevant at this port
-//                        //                        if checkForValueTypeHere {
-//                        //                            let valueType = portValue.value_type
-//                        //                            nativePatchValueTypeSettings
-//                        //                                .updateValue(.init(node_id: patchNodeData.id,
-//                        //                                                   value_type: valueType),
-//                        //                                             forKey: patchNodeData.id)
-//                        //                        }
-//                        
-//                    case .stateRef(let ref):
-//                        return .ref(ref)
-//                        //                        // Check for edges here
-//                        //                        if let upstreamData = varNameOutputPortMap.get(ref) {
-//                        //                            try SwiftParserInitializerType.subscriptRef(upstreamData)
-//                        //                                .parseStitchActions(varName: varName,
-//                        //                                                    varNameIdMap: varNameIdMap,
-//                        //                                                    varNameOutputPortMap: varNameOutputPortMap,
-//                        //                                                    customPatchInputValues: &customPatchInputValues, varNamePatchNodeRefMap: varNamePatchNodeRefMap,
-//                        ////                                                        stateVarToInteractionOutputsMap: stateVarToInteractionOutputsMap,
-//                        //                                                    nativePatchNodes: nativePatchNodes,
-//                        //                                                    patchConnections: &patchConnections,
-//                        //                                                    viewStatePatchConnections: &viewStatePatchConnections,
-//                        //                                                    nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-//                        //                                                    preprocessedJSNodes: &preprocessedJSNodes,
-//                        //                                                    varNameJsFnMap: &varNameJsFnMap,
-//                        //                                                    subscriptParentInfo: .init(node_id: patchNodeData.id,
-//                        //                                                                               port_index: portIndex))
-//                        //                        } else {
-//                        //                            fatalErrorIfDebug("Expected to find subscript data")
-//                        //                        }
-//                    }
-//                }
-//                
-//                return .normal(.portValuesInit(portDataList))
             }
         }
         
@@ -394,13 +343,6 @@ extension Patch {
                 
                 return upstreamOutputValue.first?.toNodeType
                 
-                // Second, check if we're reading state for some interaction
-                //                    else if let upstreamInteractionData = stateVarToInteractionOutputsMap
-                //                            .get(refName) {
-                //                        upstreamCoordinate = .init(node_id: upstreamInteractionData.node_id,
-                //                                                   port_index: upstreamInteractionData.port_index)
-                //                    }
-                
             case .values(let values):
                 return values.first?.toNodeType
             }
@@ -408,250 +350,4 @@ extension Patch {
         
         return nil
     }
-}
-
-extension SwiftParserInitializerType {
-//    /// Creates custom input values, edges, and custom node types for patch graph data.
-//    /// Optional parameter `downstreamNodeIdCaller` called in scenarios where recursion is used.
-//    @MainActor
-//    func parseStitchActions(varName: String,
-//                            varNameIdMap: [String : String],
-//                            varNameOutputPortMap: [String : SwiftParserSubscript],
-//    customPatchInputValues: inout [CurrentAIGraphData.CustomPatchInputValue],
-//                            varNamePatchNodeRefMap: [String : String],
-////                            stateVarToInteractionOutputsMap: [String: CurrentAIGraphData.NodeIndexedCoordinate],
-//                            nativePatchNodes: [String: CurrentAIGraphData.PatchNode],
-//                            patchConnections: inout [CurrentAIGraphData.PatchConnection],
-//                            viewStatePatchConnections: inout [String : AIGraphData_V0.NodeIndexedCoordinate],
-//                            nativePatchValueTypeSettings: inout [String: CurrentAIGraphData.NativePatchNodeValueTypeSetting],
-//                            preprocessedJSNodes: inout [CurrentAIGraphData.PreprocessedJSPatchNode],
-//                            varNameJsFnMap: inout [String: String],
-//                            subscriptParentInfo: AIGraphData_V0.NodeIndexedCoordinate? = nil) throws {
-//        switch self {
-//        case .patchNode(let patchNodeData):
-//            // Marks an input port to check for a custom node type, if supported by this node
-//            guard let patch = nativePatchNodes.get(patchNodeData.id)?.node_name.value.patch else {
-//                fatalErrorIfDebug()
-//                return
-//            }
-//            
-//            let nodeValueTypeDynamicPortIndices = patch.nonStaticTypedInputPorts ?? .init()
-//            
-//            for (portIndex, arg) in patchNodeData.args.enumerated() {
-//                // Determine a custom node value type if this node supports value types and no value has yet been set here
-//                let checkForValueTypeHere = nodeValueTypeDynamicPortIndices.contains(portIndex) && !nativePatchValueTypeSettings.keys.contains(patchNodeData.id)
-//                
-//                switch arg {
-//                case .binding(let refName):
-//                    // Get edge data
-//                    let upstreamCoordinate: AIGraphData_V0.NodeIndexedCoordinate
-//                    
-//                    // First check for some other patch's outputs
-//                    if let upstreamRefData = varNameOutputPortMap.get(refName) {
-//                        upstreamCoordinate = SwiftParserPatchData
-//                            .derivePatchUpstreamCoordinate(upstreamRefData: upstreamRefData,
-//                                                           varNameIdMap: varNameIdMap)
-//                    }
-//                    
-//                    // Second, check if we're reading state for some interaction
-////                    else if let upstreamInteractionData = stateVarToInteractionOutputsMap
-////                            .get(refName) {
-////                        upstreamCoordinate = .init(node_id: upstreamInteractionData.node_id,
-////                                                   port_index: upstreamInteractionData.port_index)
-////                    }
-//                    
-//                    else {
-//                        fatalError()
-//                        continue
-//                    }
-//                    
-//                    SwiftParserPatchData
-//                        .processIncomingConnectionData(upstreamCoordinate: upstreamCoordinate,
-//                                                       downstreamCoordinate: .init(node_id: patchNodeData.id,
-//                                                                                   port_index: portIndex),
-//                                                       nativePatchNodes: nativePatchNodes,
-//                                                       nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-//                                                       patchConnections: &patchConnections)
-//                    
-//                case .value(let argType):
-//                    let portDataList = try argType.derivePortValues()
-//                    
-//                    for portData in portDataList {
-//                        switch portData {
-//                        case .value(let portValue):
-//                            customPatchInputValues.append(
-//                                .init(patch_input_coordinate: .init(
-//                                    node_id: patchNodeData.id,
-//                                    port_index: portIndex),
-//                                      value: portValue.value,
-//                                      value_type: portValue.value_type)
-//                            )
-//                            
-//                            // Update node's custom value type if relevant at this port
-//                            if checkForValueTypeHere {
-//                                let valueType = portValue.value_type
-//                                nativePatchValueTypeSettings
-//                                    .updateValue(.init(node_id: patchNodeData.id,
-//                                                       value_type: valueType),
-//                                                 forKey: patchNodeData.id)
-//                            }
-//                            
-//                        case .stateRef(let ref):
-//                            // Check for edges here
-//                            if let upstreamData = varNameOutputPortMap.get(ref) {
-//                                try SwiftParserInitializerType.subscriptRef(upstreamData)
-//                                    .parseStitchActions(varName: varName,
-//                                                        varNameIdMap: varNameIdMap,
-//                                                        varNameOutputPortMap: varNameOutputPortMap,
-//                                                        customPatchInputValues: &customPatchInputValues, varNamePatchNodeRefMap: varNamePatchNodeRefMap,
-////                                                        stateVarToInteractionOutputsMap: stateVarToInteractionOutputsMap,
-//                                                        nativePatchNodes: nativePatchNodes,
-//                                                        patchConnections: &patchConnections,
-//                                                        viewStatePatchConnections: &viewStatePatchConnections,
-//                                                        nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-//                                                        preprocessedJSNodes: &preprocessedJSNodes,
-//                                                        varNameJsFnMap: &varNameJsFnMap,
-//                                                        subscriptParentInfo: .init(node_id: patchNodeData.id,
-//                                                                                   port_index: portIndex))
-//                            } else {
-//                                fatalErrorIfDebug("Expected to find subscript data")
-//                            }
-//                        }
-//                    }
-//                    
-//                case .subscriptRef(let subscriptRef):
-//                    // Recursively call subscript data
-//                    let subscriptInitializer = SwiftParserInitializerType.subscriptRef(subscriptRef)
-//                    try subscriptInitializer
-//                        .parseStitchActions(varName: varName,
-//                                            varNameIdMap: varNameIdMap,
-//                                            varNameOutputPortMap: varNameOutputPortMap,
-//                                            customPatchInputValues: &customPatchInputValues,
-//                                            varNamePatchNodeRefMap: varNamePatchNodeRefMap,
-////                                            stateVarToInteractionOutputsMap: stateVarToInteractionOutputsMap,
-//                                            nativePatchNodes: nativePatchNodes,
-//                                            patchConnections: &patchConnections,
-//                                            viewStatePatchConnections: &viewStatePatchConnections,
-//                                            nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-//                                            preprocessedJSNodes: &preprocessedJSNodes,
-//                                            varNameJsFnMap: &varNameJsFnMap,
-//                                            subscriptParentInfo: .init(node_id: patchNodeData.id,
-//                                                                       port_index: portIndex))
-//                }
-//            }
-//            
-//        case .stateMutation(let mutationData):
-//            let subscriptData: SwiftParserSubscript
-//            
-//            // Find subscript data which must exist for view state mutation
-//            switch mutationData {
-//            case .subscriptRef(let _subscriptData):
-//                subscriptData = _subscriptData
-//                
-//            case .patchNodeRef:
-//                // Check if we need this
-//                fatalError()
-//                
-//            case .declrRef(let ref):
-//                guard let refData = varNameOutputPortMap.get(ref) else {
-//                    throw SwiftUISyntaxError.unexpectedStateMutatorFound(mutationData)
-//                }
-//                
-//                subscriptData = refData
-//                
-//            default:
-//                return
-//            }
-//            
-//            // Track upstream patch coordinate to some TBD layer input
-//            let usptreamCoordinate = SwiftParserPatchData
-//                .derivePatchUpstreamCoordinate(upstreamRefData: subscriptData,
-//                                               varNameIdMap: varNameIdMap)
-//            
-//            viewStatePatchConnections.updateValue(usptreamCoordinate,
-//                                                  forKey: varName)
-//            
-//        case .subscriptRef(let subscriptData):
-//            switch subscriptData.subscriptType {
-//            case .patchNode(let patchNodeData):
-//                let initializerData = SwiftParserInitializerType.patchNode(patchNodeData)
-//                
-//                // Recursively parse patch node data
-//                try initializerData
-//                    .parseStitchActions(varName: varName,
-//                                        varNameIdMap: varNameIdMap,
-//                                        varNameOutputPortMap: varNameOutputPortMap,
-//                                        customPatchInputValues: &customPatchInputValues,
-//                                        varNamePatchNodeRefMap: varNamePatchNodeRefMap,
-////                                        stateVarToInteractionOutputsMap: stateVarToInteractionOutputsMap,
-//                                        nativePatchNodes: nativePatchNodes,
-//                                        patchConnections: &patchConnections,
-//                                        viewStatePatchConnections: &viewStatePatchConnections,
-//                                        nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-//                                        preprocessedJSNodes: &preprocessedJSNodes,
-//                                        varNameJsFnMap: &varNameJsFnMap,)
-//                
-//            case .ref(let refName):
-//                // Get edge data
-//                guard let destNodeId = varNameIdMap.get(refName),
-//                      let destCoordinate = subscriptParentInfo else {
-//                    // Some refs can be ignored
-//                    return
-//                }
-//                
-//                SwiftParserPatchData
-//                    .processIncomingConnectionData(
-//                        upstreamCoordinate: .init(node_id: destNodeId,
-//                                                  port_index: subscriptData.portIndex),
-//                        downstreamCoordinate: destCoordinate,
-//                        nativePatchNodes: nativePatchNodes,
-//                        nativePatchValueTypeSettings: &nativePatchValueTypeSettings,
-//                        patchConnections: &patchConnections)
-//            }
-//        
-//        case .jsNodeScript(let script):
-//            // Must reuse ID
-//            guard let varNameForJsFn = varNameJsFnMap.get(varName),
-//                  let id = varNameIdMap.get(varNameForJsFn) else {
-//                // Ignore js node if nothing uses it
-//                break
-//            }
-//            
-//            // Create JS node
-//            let newJSNode = AIGraphData_V0
-//                .PreprocessedJSPatchNode(node_id: id,
-//                                         funcName: varName,
-//                                         sourceCode: script)
-//            preprocessedJSNodes.append(newJSNode)
-//            
-//        case .viewBuilder, .patchNodeRef, .declrRef, .arraySyntax:
-//            return
-//        }
-//    }
-}
-
-extension AIGraphData_V0.NodeIndexedCoordinate {
-    /// Determines a custom value type of some node given data from its upstream node connection.
-//    @MainActor
-//    func determineOutputNodeValueType(nativePatchNodes: [String: CurrentAIGraphData.PatchNode],
-//                                      nativePatchValueTypeSettings: [String: CurrentAIGraphData.NativePatchNodeValueTypeSetting]) -> StitchAIPortValue_V1.NodeType? {
-//        guard let upstreamNode = nativePatchNodes.get(self.node_id) else {
-//
-//            
-//            return nil
-//        }
-//        
-//        let upstreamNodeType = nativePatchValueTypeSettings.get(self.node_id)?.value_type.value
-//        guard let upstreamPatch = upstreamNode.node_name.value.patch else {
-//            fatalErrorIfDebug()
-//            return nil
-//        }
-//        
-//        guard let upstreamOutput = upstreamPatch.graphNode?.rowDefinitions(for: upstreamNodeType).outputs[safe: self.port_index] else {
-//            fatalErrorIfDebug()
-//            return nil
-//        }
-//        
-//        return upstreamOutput.value.nodeType
-//    }
 }
